+++ tests/framework/db/mssql/schema/SchemaTest.php
@@ -26,28 +26,28 @@ final class SchemaTest extends \yiiunit\framework\db\schema\AbstractSchema
     }
 
     /**
-     * @dataProvider \yiiunit\framework\db\mssql\provider\SchemaProvider::resetSequence
+     * @dataProvider \yiiunit\framework\db\mssql\provider\SchemaProvider::resetAutoIncrementPK
      */
-    public function testResetSequence(
+    public function testResetAutoIncrementPK(
         string $tableName,
         array $insertRows,
         array $expectedIds,
         int|null $value = null
     ): void {
-        parent::testResetSequence($tableName, $insertRows, $expectedIds, $value);
+        parent::testResetAutoIncrementPK($tableName, $insertRows, $expectedIds, $value);
     }
 
-    public function testResetSequenceWithTableNotPrimaryKey(): void
+    public function testResetAutoIncrementPKWithTableNotPrimaryKey(): void
     {
         $this->columnsSchema = [
             'id' => 'INT',
             'name' => 'NVARCHAR(128)',
         ];
 
-        parent::testResetSequenceWithTableNotPrimaryKey();
+        parent::testResetAutoIncrementPKWithTableNotPrimaryKey();
     }
 
-    public function testResetSequenceWithTablePrimaryKeyComposite(): void
+    public function testResetAutoIncrementPKWithTablePrimaryKeyComposite(): void
     {
         $this->columnsSchema = [
             'id' => 'INT IDENTITY',
@@ -56,6 +56,6 @@ final class SchemaTest extends \yiiunit\framework\db\schema\AbstractSchema
             'PRIMARY KEY (id, user_id)',
         ];
 
-        parent::testResetSequenceWithTablePrimaryKeyComposite();
+        parent::testResetAutoIncrementPKWithTablePrimaryKeyComposite();
     }
 }